+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Allow `GenServer` start options to be provided when starting event handlers and process managers ([#398](https://github.com/commanded/commanded/pull/398)).
 - Add `hibernate_after` option to application config ([#399](https://github.com/commanded/commanded/pull/399)).
 - Add support for providing adapter-specific event store subscription options ([#391](https://github.com/commanded/commanded/pull/391)).
+- Support custom state for event handlers ([#400](https://github.com/commanded/commanded/pull/400)).
 
 ## v1.1.1